Improved shared functions
